refactor: drop unused GitHub review helpers
- Delete unused GitHub, quota, audit, history, and intent helpers
- Drop unused submitReview constants and repair prompt suffixes
- Collapse askSafety re-exports, Context7 parsing, and the file-list
  wrap
- Trim GITHUB_API_TOOL_NAMES to live publish and submit tools
- Point ack progress at ACTIVE_WORK_STATUSES
