+++ CONTEXT.md
@@ -71,7 +71,7 @@ This file is **domain language only** — not a specification of how the system 
 - **Review summary comment** — The completed form of a review progress comment: navigation and overview gates, not duplicated finding bodies from inline review threads. Owns "what's risky" (gates, findings table) and the aggregate **agent fix prompt** accordion below the table; never carries a per-file walkthrough or review map. When a published description agent block includes a **review map**, the summary links to that map instead of duplicating it; short omit-mode PRs get no link.
 - **Review overview (`prCharacter`)** — Free-form Note on the review summary comment. Authored at synthesis via `publish_summary`. Uses the same **body scale** map as description writing policy (brief / standard / detailed from workspace size stats) plus ASD-STE100 short active sentences. States what the PR changes and review stakes; never inventories files, restates the findings table, or reports specialist-lane process.
 - **Review size (`size`)** — T-shirt scale (`XS`, `S`, `M`, `L`, `XL`, `XXL`) for the magnitude of the change set, not code quality or review cost. Model-authored at synthesis. Shown as the Size row on the **review summary comment** and synced onto the PR as the managed `size:<TOKEN>` label when `FEATURE_REVIEW_LABELS` is on.
-- **Finding confidence** — Per-finding certainty level (`c1`–`c5`) shown beside severity in the review summary comment; expresses how sure the reviewer is about one finding, nothing else. Prompts ask models to drop scored findings below the configured minimum (`REVIEW_MIN_CONFIDENCE`, default `1`).
+- **Finding confidence** — Per-finding certainty level (`c1`–`c5`) shown beside severity in the review summary comment; expresses how sure the reviewer is about one finding, nothing else. The specialist finding-field contract tells the model to drop anything it would mark `1`.
 - **CI summary** — Gate row on the **review progress comment** stub and completed **review summary comment**. Server fetches external check runs, legacy commit statuses, and (when failing) one condensed, redacted, size-bounded CI context for the PR head (Actions job logs, or check output if logs are unavailable; excluding PR Agent’s own check), then a small LLM turn authors structured fields (`headline`, per-failure `reason` / `fixHint`). Failing-head annotation and job-log fetches run concurrently within `REVIEW_CI_SUMMARY_FETCH_CONCURRENCY`; workflow runs are filtered to the reviewed head before any job listing, and raw job logs are bounded to a failure-preserving intake window before condensation. Server owns status facts, check-run completeness, wait/poll, redaction, schema validation, and HTML rendering; the model must not emit the table markup. Ack uses a lightweight non-LLM snapshot only. When CI is still pending at publish, the row stays pending; a later `workflow_run` or `check_suite` completed webhook refreshes the CI cell on the matching summary without a full re-review. A refresh that arrives while a review is still queued or running is retained and retried on the same lane for that head SHA until it can patch or the attempt cap is exhausted. An incomplete check-run page (pagination cap) is **unavailable**, not passing or none, and does not use the grant-Checks copy. Known failures in that partial view stay failing; the headline notes the partial view. A later complete poll or refresh can replace the snapshot. Missing Checks permission shows a grant-Checks row; missing Actions permission on a red head keeps the failure row and adds a grant-Actions note. Caps live in code constants (`REVIEW_CI_SUMMARY_*`). See [ADR 0018](docs/adr/0018-llm-authored-ci-summary.md).
 - **Prose contract** — Changed Markdown that defines behavior (agent instructions, skill definitions, configuration docs) rather than merely describing it. The review orchestrator checks prose contracts for internal consistency: counts vs listed items, cross-references that resolve, command examples that are neither broken nor destructive, stated defaults vs the rest of the diff. Contradictions are ordinary findings, not style nits.
 - **Truncated change set** — File listing for an orchestrated review run where some changed files are omitted due to configured caps; the run continues with explicit truncation metadata.
