+++ src/agentWork/types.ts
@@ -9,16 +9,6 @@ export type WorkType = "review" | "ask" | "description" | "triage" | "verificati
 export type WorkStatus = "queued" | "running" | "superseded" | "cancelled" | "completed" | "failed";
 
 export const ACTIVE_WORK_STATUSES = ["queued", "running"] as const satisfies readonly WorkStatus[];
-export const TERMINAL_WORK_STATUSES = [
-  "superseded",
-  "cancelled",
-  "completed",
-  "failed",
-] as const satisfies readonly WorkStatus[];
-
-export function isTerminalWorkStatus(status: string): boolean {
-  return (TERMINAL_WORK_STATUSES as readonly string[]).includes(status);
-}
 
 export type WebhookHeaders = {
   readonly delivery?: string;
