+++ test/askIntake.test.ts
@@ -2,9 +2,14 @@ import { describe, expect, it, vi } from "vitest";
 import type { PoolClient } from "pg";
 import type { PgBoss } from "pg-boss";
 import { promoteAskFromWebhookEvent } from "../src/agentWork/intake/askIntake.js";
-import { ACK_QUEUE, ASK_QUEUE, ASK_THROTTLED_BODY, ASK_USAGE_HINT } from "../src/settings/index.js";
+import {
+  ACK_QUEUE,
+  ASK_QUEUE,
+  ASK_THROTTLED_BODY,
+  ASK_USAGE_HINT,
+  MAX_ASK_QUESTION_CHARS,
+} from "../src/settings/index.js";
 import { ASK_QUESTION_TOO_LONG_HINT } from "../src/commands/parseAskQuestion.js";
-import { MAX_ASK_QUESTION_CHARS } from "../src/agent/ask/askSafety.js";
 import { defaultAskQuotaConfig } from "../src/agentWork/askQuota.js";
 
 function baseInput(overrides: Record<string, unknown> = {}) {
