+++ test/schedulerAsk.test.ts
@@ -6,9 +6,13 @@ import type { Pool, PoolClient } from "pg";
 import type { PgBoss } from "pg-boss";
 import { createOperationLogger } from "../src/evlog.js";
 import { makeAgentWorkScheduler } from "../src/agentWork/scheduler.js";
-import { MAX_ASK_QUESTION_CHARS } from "../src/agent/ask/askSafety.js";
+import {
+  ACK_QUEUE,
+  ASK_QUEUE,
+  ASK_USAGE_HINT,
+  MAX_ASK_QUESTION_CHARS,
+} from "../src/settings/index.js";
 import { ASK_QUESTION_TOO_LONG_HINT } from "../src/commands/parseAskQuestion.js";
-import { ACK_QUEUE, ASK_QUEUE, ASK_USAGE_HINT } from "../src/settings/index.js";
 import * as postgres from "../src/db/postgres.js";
 
 function makeSlashInput(body: string) {
